Battle logic commit
This commit contains a basic battle function gameplay loop, which can be used in the future for all typical combat. Next I would like to import a random library to add some RNG to the game and make everything a bit easier to read.

Then I will populate the first locations with enemies and other functions.